more DEF_PAGER
For Unix, add internal vs external pager choice to #version.
Others always use internal pagination so don't need to see that.

Also remove obsolete "command completion" option.  I don't know
when it became unconditional but that was long enough ago to be
absent from the git log and from second cvs log included in that.

And streamline the RNG seed stuff a bit.