out-of-bounds access in farlook pr592
https://github.com/NetHack/NetHack/pull/592 comment states:
"In commit 35d0149, most of the instances of xdir and ydir here were
changed to u.dx and u.dy, but not all of them. The remaining ones are
out-of-bounds on xdir and ydir, because i is always set to 12 from an
earlier loop and is no longer involved in handling user input. They
should be u.dx and u.dy like the rest."