implement github issue #727 - item-action '#tip'
Issue reported by Meklon2007:  picking a container from inventory
didn't offer #tip as a possible use of the item.
Implemented in commit f7d3f7006d65a216e23ec3bae278ed47e7af4a19
but I forgot to close the issue, so this adds a fixes entry for
it in order to do that.

Closes #727