more shop wall repair
Commit d440b34ebfb754a61227bbe09fa6479cc2d74703 took the 'catchup'
argument away from repair_damage() and calculated it in the routine.
Commit a8447afcdeae4076ece03632cd48e3e83e9442e7 put the argument
back but neglected to remove calculating and assigning it, making
the passed value be ignored.  Take that away, finishing a8447afcde.

That affected shop repair messages but wasn't enough to prevent a
shop wall repair display glitch.  This seems to make things work
properly but is little iffy.